+++ packages/eslint-config-smarthr/CHANGELOG.md
@@ -2,6 +2,55 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [14.11.0](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.10.0...eslint-config-smarthr-v14.11.0) (2026-07-22)
+
+
+### Features
+
+* update ESLint rule severities in eslint-config-smarthr ([#1476](https://github.com/kufu/tamatebako/issues/1476)) ([5213e9f](https://github.com/kufu/tamatebako/commit/5213e9f2b1659bd2b3bfe7cdb118a2ed001aa072))
+
+## [14.10.0](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.9.1...eslint-config-smarthr-v14.10.0) (2026-07-14)
+
+
+### Features
+
+* eslint-plugin-smarthr 6.23.0への更新（eslint-config-smarthr） ([#1457](https://github.com/kufu/tamatebako/issues/1457)) ([c7c3f9a](https://github.com/kufu/tamatebako/commit/c7c3f9af605a1b21f128bfdebc976f5d99778e64))
+
+## [14.9.1](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.9.0...eslint-config-smarthr-v14.9.1) (2026-07-03)
+
+
+### Bug Fixes
+
+* update eslint-plugin-smarthr to v6.22.1 in eslint-config-smarthr ([#1448](https://github.com/kufu/tamatebako/issues/1448)) ([3996dd6](https://github.com/kufu/tamatebako/commit/3996dd619736e00f067f5a296218bd97be657d1e))
+
+## [14.9.0](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.8.2...eslint-config-smarthr-v14.9.0) (2026-07-02)
+
+
+### Features
+
+* eslint-plugin-smarthrをv6.22.0に更新してbest-practice-for-reduce-redundant-callsを導入 ([#1443](https://github.com/kufu/tamatebako/issues/1443)) ([db68985](https://github.com/kufu/tamatebako/commit/db689859de1c2582e3f670665c778e048f3d1336))
+
+## [14.8.2](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.8.1...eslint-config-smarthr-v14.8.2) (2026-06-30)
+
+
+### Bug Fixes
+
+* eslint-config-smarthrのeslint-plugin-smarthrをv6.21.2に更新 ([#1431](https://github.com/kufu/tamatebako/issues/1431)) ([9382893](https://github.com/kufu/tamatebako/commit/93828931f5de888b72353ff493c2eaf09914d94d))
+
+## [14.8.1](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.8.0...eslint-config-smarthr-v14.8.1) (2026-06-22)
+
+
+### Bug Fixes
+
+* **eslint-config-smarthr:** use HTTPS repository metadata ([#1403](https://github.com/kufu/tamatebako/issues/1403)) ([9f439c8](https://github.com/kufu/tamatebako/commit/9f439c82b4b6908a1a2f092a87bb7709f45e3049))
+
+## [14.8.0](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.7.0...eslint-config-smarthr-v14.8.0) (2026-06-22)
+
+
+### Features
+
+* eslint-plugin-smarthrを6.21.0に更新（eslint-config-smarthr） ([#1409](https://github.com/kufu/tamatebako/issues/1409)) ([08be327](https://github.com/kufu/tamatebako/commit/08be3274393797cc87cf862fa68f4b98c0b089c8))
+
 ## [14.7.0](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.6.0...eslint-config-smarthr-v14.7.0) (2026-06-16)